requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0

# Code analysis
tree-sitter>=0.20.0
astroid>=2.15.0
pygments>=2.15.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration management
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
colorama>=0.4.6

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Evaluation metrics
scikit-learn>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2
